+++ packages/build-utils/src/types.ts
@@ -411,11 +411,13 @@ export interface Builder {
   config?: Config;
 }
 
+export type MaxDuration = number | 'max';
+
 export interface BuilderFunctions {
   [key: string]: {
     architecture?: LambdaArchitecture;
     memory?: number;
-    maxDuration?: number;
+    maxDuration?: MaxDuration;
     regions?: string[];
     functionFailoverRegions?: string[];
     runtime?: string;
@@ -769,7 +771,7 @@ export interface ExperimentalServiceConfig {
 
   /** Lambda config */
   memory?: number;
-  maxDuration?: number;
+  maxDuration?: MaxDuration;
   includeFiles?: string | string[];
   excludeFiles?: string | string[];