+++ packages/cervel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/cervel
 
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/backends@0.0.45
+
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/backends@0.0.44
+
 ## 0.0.30
 
 ### Patch Changes